dev_requirements.txt
-e .

requests
pytest
mock
PyYAML
honcho
portend
webtest
bumpversion

Flask-Script
Flask-Migrate
